# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ContentMap {[key: string]: string}

export type ContentUpdateEntries = {[key: string]: string};

export interface ContentUpdate {
  entries: ContentUpdateEntries;
}

export interface AdminCredentials {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface ApiMessage {
  error: string;
}

export interface HealthStatus {
  status: string;
}