# lib/api-zod/src/generated/types/contentUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentUpdateEntries } from './contentUpdateEntries';

export interface ContentUpdate {
  entries: ContentUpdateEntries;
}